Render identifier constants special, fix slot styles
