+++ src/components/block/slot/ArgumentSlot.tsx
@@ -1,5 +1,5 @@
 import { BlockLabels } from "../block-aria-labels.ts";
-import { Box, type BoxProps } from "@mui/material";
+import { Box } from "@mui/material";
 import { Block, type BlockProps } from "../Block.tsx";
 import { useDroppable } from "@dnd-kit/core";
 import { Draggable } from "../../../common/dnd-wrappers/Draggable.tsx";
@@ -14,27 +14,6 @@ export interface ArgumentSlotProps {
 }
 
 export const ArgumentSlotPrefix = "argument slot of ";
-const FakeSlotStyles: BoxProps = {
-  bgcolor: "grey",
-  position: "relative",
-  sx: {
-    isolation: "isolate",
-    "&::before": {
-      position: "absolute",
-      top: "50%",
-      left: "50%",
-      transform: "translate(-50%, -50%)",
-      content: '"+"',
-      textAlign: "center",
-      fontSize: "1rem",
-      fontWeight: "bold",
-      color: "text.primary",
-      opacity: 0.5,
-      zIndex: -1,
-      userSelect: "none",
-    },
-  },
-};
 
 export function ArgumentSlot({
   idSuffix,
@@ -51,7 +30,7 @@ export function ArgumentSlot({
   );
   return (
     <Box
-      {...ArgumentSlotStyles}
+      {...ArgumentSlotStyles(fake)}
       aria-label={BlockLabels.ArgumentSlot}
       ref={setNodeRef}
       {...(isOver
@@ -59,7 +38,6 @@ export function ArgumentSlot({
             boxShadow: "inset 0 0 0 0.25em lightgreen",
           }
         : null)}
-      {...(fake ? FakeSlotStyles : null)}
     >
       {ChildBlock ? (
         <Draggable id={ChildBlock.props.id}>{ChildBlock}</Draggable>
