+++ src/components/block/slot/argument-slot-styles.ts
@@ -1,6 +1,29 @@
 import type { BoxProps } from "@mui/material";
 
-export const ArgumentSlotStyles: BoxProps = {
+const FakeSlotStyles: BoxProps = {
+  bgcolor: "grey",
+  position: "relative",
+  sx: {
+    isolation: "isolate",
+    "&::before": {
+      position: "absolute",
+      top: "50%",
+      left: "50%",
+      transform: "translate(-50%, -50%)",
+      content: '"+"',
+      textAlign: "center",
+      fontSize: "1rem",
+      fontWeight: "bold",
+      color: "text.primary",
+      opacity: 0.5,
+      zIndex: -1,
+      userSelect: "none",
+    },
+  },
+};
+export const ArgumentSlotStyles: (fake: boolean | undefined) => BoxProps = (
+  fake,
+) => ({
   minHeight: "0.5em",
   minWidth: "1.5em",
   width: "fit-content",
@@ -8,4 +31,5 @@ export const ArgumentSlotStyles: BoxProps = {
   bgcolor: "white",
   padding: "0.4em",
   borderRadius: "0.5em",
-};
+  ...(fake ? FakeSlotStyles : null),
+});
